refactor(060): move puzzle sample out of domain
