+++ app/src/main/java/org/cescfe/numpairs/ui/screen/GameViewModel.kt
@@ -4,15 +4,15 @@ import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import org.cescfe.numpairs.initialPuzzle as defaultInitialPuzzle
 import org.cescfe.numpairs.domain.puzzle.Board
 import org.cescfe.numpairs.domain.puzzle.OperandSlot
 import org.cescfe.numpairs.domain.puzzle.Operator
 import org.cescfe.numpairs.domain.puzzle.Puzzle
-import org.cescfe.numpairs.domain.puzzle.PuzzleSamples
 import org.cescfe.numpairs.domain.puzzle.StripItem
 import org.cescfe.numpairs.domain.puzzle.operandSelectionHintsFor
 
-class GameViewModel(initialPuzzle: Puzzle = PuzzleSamples.prototype) : ViewModel() {
+class GameViewModel(initialPuzzle: Puzzle = defaultInitialPuzzle) : ViewModel() {
     private var puzzle: Puzzle = initialPuzzle
     private var stripItemEntryDialogIndex: Int? = null
     private var tileOperatorSelectionDialogIndex: Int? = null
