+++ app/src/test/java/org/cescfe/numpairs/InitialPuzzleTest.kt
@@ -1,13 +1,16 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs
 
+import org.cescfe.numpairs.domain.puzzle.Expression
+import org.cescfe.numpairs.domain.puzzle.Operator
+import org.cescfe.numpairs.domain.puzzle.StripItem
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Test
 
-class PuzzleSamplesTest {
+class InitialPuzzleTest {
     @Test
-    fun prototype_strip_starts_with_known_and_hidden_items() {
-        val stripItems = PuzzleSamples.prototype.strip.items
+    fun initial_strip_starts_with_known_and_hidden_items() {
+        val stripItems = initialPuzzle.strip.items
 
         assertEquals(5, stripItems.count { it == StripItem.Hidden })
         assertEquals(3, stripItems.count { it is StripItem.Known })
@@ -15,8 +18,8 @@ class PuzzleSamplesTest {
     }
 
     @Test
-    fun prototype_tiles_start_with_hidden_expressions() {
-        val boardTiles = PuzzleSamples.prototype.board.tiles
+    fun initial_tiles_start_with_hidden_expressions() {
+        val boardTiles = initialPuzzle.board.tiles
 
         boardTiles.forEach { tile ->
             assertEquals(Expression.Operand.Hidden, tile.expression.leftOperand)
@@ -26,8 +29,8 @@ class PuzzleSamplesTest {
     }
 
     @Test
-    fun prototype_tiles_keep_their_expected_results() {
-        val boardTiles = PuzzleSamples.prototype.board.tiles
+    fun initial_tiles_keep_their_expected_results() {
+        val boardTiles = initialPuzzle.board.tiles
 
         assertEquals(listOf(223, 222, 52, 100, 31, 150, 35, 250), boardTiles.map { it.result })
     }
