+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/support/DomainPuzzleMothers.kt
@@ -1,11 +1,11 @@
 package org.cescfe.numpairs.domain.puzzle.support
 
+import org.cescfe.numpairs.initialPuzzle
 import org.cescfe.numpairs.domain.puzzle.Puzzle
-import org.cescfe.numpairs.domain.puzzle.PuzzleSamples
 import org.cescfe.numpairs.domain.puzzle.Strip
 import org.cescfe.numpairs.domain.puzzle.StripItem
 
-fun puzzleWithRepeatedSixes(): Puzzle = PuzzleSamples.prototype.copy(
+fun puzzleWithRepeatedSixes(): Puzzle = initialPuzzle.copy(
     strip = Strip.fromItems(
         items = listOf(
             StripItem.Known(6),
